Initial Styling on price calc
Installed mui lab
